Modularize web/app.js into ES modules
Split the 942-line monolith into focused ES modules:
- utils.js: formatting helpers, validation, HTML escaping
- parser.js: CSV record, block, and transaction parsing
- dataset.js: data model construction and address profiling
- renderer.js: DOM rendering for all views and components
- app.js: entry point and state management (~160 lines)

Update index.html to type="module" and Makefile ui-build to copy
all JS files into web/dist/.

UI smoke tests pass; no behavioral changes.
